+++ frontend/src/routes/ProtectedRoute.tsx
@@ -1,9 +1,7 @@
 import { useKeycloak } from "@react-keycloak/web";
-import { FC, useEffect } from "react";
-interface Props {
-  component: FC;
-}
-export const ProtectedRoute = ({ component }: { component: JSX.Element }) => {
+import { FC, ReactChild, ReactNode, useEffect } from "react";
+
+const ProtectedRoute = ({ children }: { children: ReactNode }) => {
   const { keycloak, initialized } = useKeycloak();
 
   useEffect(() => {
@@ -16,6 +14,7 @@ export const ProtectedRoute = ({ component }: { component: JSX.Element }) => {
     return <div>Loading...</div>;
   }
 
-  return component;
+  return <>{children}</>; // Wrapping children with empty fragment to avoid wrapper div warnings
 };
+
 export default ProtectedRoute;