+++ frontend/src/routes/subRoutes/adminRoutes.tsx
@@ -13,7 +13,16 @@ import { Glossary } from "pages/Admin/Glossary";
  * Routes for Admin pages.
  */
 const routes = [
-  <Route key="admin" path="/admin" element={<ProtectedRoute component={<Admin />} />}>
+  <Route
+    key="admin"
+    path="/admin"
+    element={
+      <ProtectedRoute>
+        <Admin />
+      </ProtectedRoute>
+    }
+  >
+    ,
     <Route key="contacts" path="contacts" element={<Contacts />} />
     <Route key="suppliers" path="suppliers" element={<Suppliers />} />
     <Route key="subcontractors" path="subcontractors" element={<Subcontractors />} />