+++ frontend/src/routes/subRoutes/projectRoutes.tsx
@@ -9,20 +9,31 @@ import { LessonsLearned } from "pages/Projects/Project/LessonsLearned";
 import { Status } from "pages/Projects/Project/Status";
 import ProtectedRoute from "routes/ProtectedRoute";
 const projectRoutes = [
-  <Route key="project" path="projects" element={<ProtectedRoute component={<Projects />} />} />,
   <Route
-    key="projectprojectId"
+    key="projects"
+    path="/projects"
+    element={
+      <ProtectedRoute>
+        <Projects />
+      </ProtectedRoute>
+    }
+  />,
+  <Route
+    key="projectId"
     path="projects/:projectId"
-    element={<ProtectedRoute component={<Project />} />}
+    element={
+      <ProtectedRoute>
+        <Project />
+      </ProtectedRoute>
+    }
   >
     ,
-    <Route index key="projectdetails" element={<ProjectDetails />} />,
-    <Route key="projectStatus" path="status" element={<Status />} />,
-    <Route key="changeRequest" path="change-request" element={<ChangeRequest />} />,
-    <Route key="projectBilling" path="billing" element={<Billing />} />,
+    <Route index key="projectdetails" element={<ProjectDetails />} />
+    <Route key="projectStatus" path="status" element={<Status />} />
+    <Route key="changeRequest" path="change-request" element={<ChangeRequest />} />
+    <Route key="projectBilling" path="billing" element={<Billing />} />
     <Route key="projectLessonsLearned" path="lessons-learned" element={<LessonsLearned />} />
-    ,
-    <Route key="projectCloseOut" path="close-out" element={<CloseOut />} />,
+    <Route key="projectCloseOut" path="close-out" element={<CloseOut />} />
   </Route>,
 ];